More const and other small changes. Restored mayor type strings, and reset some options to default to original.
git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@711 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gfx/spritesys/FacedSprite.cpp
@@ -79,7 +79,7 @@ FacedSprite::~FacedSprite()
 
 
 
-void FacedSprite::Import(uint16 nframes, char *imageFiles[k_NUM_FACINGS][k_MAX_NAMES], char *shadowFiles[k_NUM_FACINGS][k_MAX_NAMES])
+void FacedSprite::Import(size_t nframes, char *imageFiles[k_NUM_FACINGS][k_MAX_NAMES], char *shadowFiles[k_NUM_FACINGS][k_MAX_NAMES])
 {
     AllocateFrameArrays(nframes);
